Use embedded configuration by default
I can't see any reason why the embedded mode isn't always preferable.
The emsdk is designed to be used as-is, and per-user customizations
don't make sense.

I'm hoping if this sicks we can just remove this options completely.

This is part of a wider plan to remove the use of the user's HOME
directory completely:
  https://github.com/emscripten-core/emscripten/issues/9543